% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dtmax.R
\name{calc_dtmax_30min}
\alias{calc_dtmax_30min}
\title{Calculate dTmax from half-hourly dT time series by six different methods}
\usage{
calc_dtmax_30min(vctr_time, vctr_dt, vctr_rs, vctr_vpd, prd_all)
}
\arguments{
\item{vctr_time}{Timestamp vector}

\item{vctr_dt}{Delta T time series}

\item{vctr_rs}{Incident short-wave radiation time series}

\item{vctr_vpd}{Vapor pressure deficit time series}

\item{prd_all}{Target period for dTmax estimation}
}
\description{
calc_dtmax_30min() calculates daily maximum temperature
  difference between granier-type sap flow probes (dTmax) by the predawn
  method, predawn slide method, moving window method, double average method,
  double regression method, and environmental dependent method.
}
\examples{
time_test <- dt_noisy$time
dt_test <- dt_noisy$dt
rs_test <- dt_noisy$rs
vpd_test <- dt_noisy$vpd

# Calculate dTmax from data
result <-
  calc_dtmax_30min(time_test, dt_test, rs_test, vpd_test,
                   prd_all = "2012-09-08/2013-09-07")
}
