% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dtmax.R
\name{calc_fd_30min}
\alias{calc_fd_30min}
\title{Calculate Fd from six different dTmax time series}
\usage{
calc_fd_30min(
  df_dtmax,
  alpha = 1.19 * 10^{     -4 },
  beta = 1.231,
  do_heartwood_correction = FALSE,
  ratio_conductive = NULL
)
}
\arguments{
\item{df_dtmax}{Dataframe outputted from calc_dtmax_30min()}

\item{alpha}{Parameter in Granier equation (m/s)}

\item{beta}{Parameter in Granier equation}

\item{do_heartwood_correction}{Logical. Do Heartwood correction described in Clearwater et al. (1999)}

\item{ratio_conductive}{Ratio of conductive versus non-conductive length}
}
\description{
calc_fd_30min() calculates sap flow rate using Granier equation
  from dTmax time series estimated by the predawn method,
  predawn slide method, moving window method, double average method,
  double regression method, and environmental dependent method.
}
\examples{
# Calculate Fd from data
data_dtmax <- dtmax
result <-
  calc_fd_30min(data_dtmax)
}
