% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dt_noisy}
\alias{dt_noisy}
\title{Raw dT time series with various artifacts}
\format{
A data frame with 17520 rows and 8 variables:
\describe{
  \item{time}{Timestamp of the measurement end timing}
  \item{dt}{Raw dT (degrees Celsius)}
  \item{p}{Precipitation (mm)}
  \item{sw_in}{Global solar radiation (W m-2)}
  \item{ta}{Air temperature (degrees Celsius)}
  \item{vpd}{Vapor pressure deficit (hPa)}
  \item{ws}{Horizontal wind speed (m s-1)}
  \item{swc}{Soil water content (m3 m-3)}
}
}
\source{
Meteorological factors and soil water content data were provided by
 Dr. Tomonori Kume [ORCiD: 0000-0001-6569-139X]
}
\usage{
data(dt_noisy)
}
\description{
Dataset consisting of the time series of a raw dT (the temperature
difference between thermal dissipation sap flow probes), meteorological
factors, and soil water content at Lambir Hills National Park, Malaysia.
Missing values are represented as -9999.
}
\keyword{datasets}
