% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dtmax}
\alias{dtmax}
\title{dTmax time series estimated by multiple methods}
\format{
A data frame with 17520 rows and 6 variables:
\describe{
  \item{time}{Timestamp of the measurement end timing}
  \item{dtmax_sp}{dTmax estimated by the SP method (degrees Celsius)}
  \item{dtmax_pd}{dTmax estimated by the PD method (degrees Celsius)}
  \item{dtmax_mw}{dTmax estimated by the MW method (degrees Celsius)}
  \item{dtmax_dr}{dTmax estimated by the DR method (degrees Celsius)}
  \item{dtmax_ed}{dTmax estimated by the ED method (degrees Celsius)}
}
}
\usage{
data(dtmax)
}
\description{
Dataset consisting of the time series of the dTmax (the maximum temperature
difference between thermal dissipation sap flow probes under zero-flow
conditions) calculated by the successive predawn (SP), daily predawn (PD),
moving window (MW), double regression (DR), and environmental dependent (ED)
methods.
}
\keyword{datasets}
