% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxfixer-package.R
\docType{package}
\name{fluxfixer-package}
\alias{fluxfixer}
\alias{fluxfixer-package}
\title{fluxfixer: A Sophisticated Post-processing Framework for Sap Flow Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a flexible framework for post-processing thermal dissipation sap flow data using statistical methods and machine learning, including anomaly correction, outlier removal, gap-filling, and sap flux density calculation. This framework can also apply to other raw time series with various artifacts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yhata86/fluxfixer}
  \item Report bugs at \url{https://github.com/yhata86/fluxfixer/issues}
}

}
\author{
\strong{Maintainer}: Yoshiaki Hata \email{yoshiakihata0806@gmail.com} (\href{https://orcid.org/0000-0001-7703-0863}{ORCID}) [copyright holder]

}
\keyword{internal}
