% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_interval}
\alias{get_interval}
\title{Obtain time interval of input timestamp vector}
\usage{
get_interval(vctr_time)
}
\arguments{
\item{vctr_time}{A timestamp vector of class POSIXct or POSIXlt. The
timestamps must be equally spaced and arranged chronologically.}
}
\value{
A numeric (minutes) indicating the time interval of the input timestamp
 vector.
}
\description{
`get_interval()` retrieves the time interval of the input
 timestamp vector and checks whether the format of the time vector is
 acceptable for the successive process.
}
\references{
Hata, Y. & Kumagai, T. (2026)
fluxfixer: An R package for producing thermal dissipation sap flow data
with high quality control. SoftwareX, 34, 102740.
\doi{10.1016/j.softx.2026.102740}
}
\author{
Yoshiaki Hata
}
\keyword{internal}
