% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_valid}
\alias{n_valid}
\title{Obtain the number of data points without missing values}
\usage{
n_valid(vctr, label_err = -9999)
}
\arguments{
\item{vctr}{A vector to be evaluated.}

\item{label_err}{A numeric value representing a missing value in the input
vector(s). Default is -9999.}
}
\value{
An integer indicating the number of the input vector elements without
 missing values.
}
\description{
`n_valid()` retrieves the number of data points without missing
  values.
}
\references{
Hata, Y. & Kumagai, T. (2026)
fluxfixer: An R package for producing thermal dissipation sap flow data
with high quality control. SoftwareX, 34, 102740.
\doi{10.1016/j.softx.2026.102740}
}
\author{
Yoshiaki Hata
}
\keyword{internal}
