% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_statistical_outlier.R
\name{remove_outlier}
\alias{remove_outlier}
\title{Remove outliers by statistical method}
\usage{
remove_outlier(
  vctr_time,
  vctr_dt,
  list_time_prd_tail,
  list_time_err_head = NULL,
  list_time_err_tail = NULL,
  list_time_drft_head = NULL,
  list_time_drft_tail = NULL,
  thres_min_dt = 3,
  thres_max_dt = 50,
  wndw_size = 48 * 15,
  min_n_wndw = 5,
  thres_z = 5,
  check_damping = TRUE,
  thres_ratio_damp = 0.5,
  filter_noise = FALSE,
  wndw_size_noise = 13,
  inv_sigma_gauss_noise = 0.01
)
}
\arguments{
\item{vctr_time}{Timestamp vector and must be continuous (no gap permitted)}

\item{vctr_dt}{Delta T (temperature difference between probes) time series}

\item{list_time_prd_tail}{List of timestamp describing sensor replacement}

\item{list_time_err_head}{List of timestamp describing specific error period started}

\item{list_time_err_tail}{List of timestamp describing specific error period ended}

\item{list_time_drft_head}{List of timestamp describing specific drift period started}

\item{list_time_drft_tail}{List of timestamp describing specific drift period ended}

\item{thres_min_dt}{Minimum threshold value of acceptable delta T}

\item{thres_max_dt}{Maximum threshold value of acceptable delta T}

\item{wndw_size}{The number of data points in a moving window for z-score normalization}

\item{min_n_wndw}{Minimum number of data points in a moving window for calculating z-score}

\item{thres_z}{Threshold absolute value of z-score for detecting outliers}

\item{check_damping}{Logical. If true, short-term damping is considered for z-score normalization.}

\item{thres_ratio_damp}{Threshold value of correcting short-term damping}

\item{filter_noise}{Logical. If true, filtering noise by gaussian window is conducted.}

\item{wndw_size_noise}{Window size of gaussian window}

\item{inv_sigma_gauss_noise}{Parameter for gaussian window}
}
\description{
remove_outlier() removes unreasonable values, corrects short-term
  drift, filtering noise, detects outliers by z-score normalization.
}
\examples{
# Remove outliers
library(lubridate)

time_test <- dt_noisy$time
dt_test <- dt_noisy$dt

prd_tail <- list(ymd_hm("2013/05/14 13:00"))
err_head <- list(ymd_hm("2013/05/14 13:00"),
                 ymd_hm("2016/12/14 12:30"))
err_tail <- list(ymd_hm("2013/05/14 13:00"),
                 ymd_hm("2016/12/14 21:00"))
drft_head <- list(ymd_hm("2013/05/14 13:30"))
drft_tail <- list(ymd_hm("2013/05/17 15:00"))

result <-
  remove_outlier(time_test, dt_test, prd_tail, err_head, err_tail,
                 drft_head, drft_tail)
}
