% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_randomforest.R
\name{rf_cv}
\alias{rf_cv}
\title{Cross varidation for random forest structure}
\usage{
rf_cv(
  df,
  name_label,
  list_min_nodesize,
  list_mtry,
  list_subsample,
  list_feature,
  rf_replace = TRUE,
  frac_train = 0.75,
  nfold = 10,
  ntree = 500,
  ran_seed = 12345,
  err = -9999
)
}
\arguments{
\item{df}{Data frame containing dT time series}

\item{name_label}{Input column names}

\item{list_min_nodesize}{Minimum number of random forest node size}

\item{list_subsample}{Subsample ratio}

\item{list_feature}{list vector of feature variables}

\item{rf_replace}{replace sampling of non-replace sampling in random forest}

\item{frac_train}{Fraction of training data set sampling}

\item{ntree}{The number of trees in random forest}

\item{ran_seed}{Random seed of random forest}

\item{err}{The label of error or missing values}

\item{list_m_try}{The number of input column candidate}

\item{rf_predict_all}{do prediction by constructed random forest}
}
\description{
rf_cv() constructs randow forest model trained from
  time series of temperature difference between granier-type sap flow probes
  for detecting outliers in the time series
}
\examples{
# Construct random forest structure
result <-
  rf_est(df, name_label, min_nodesize, m_try, subsample, list_feature,
         rf_replace = TRUE, rf_predict_all = FALSE,
         frac_train = 0.75, ran_seed = 12345, ntree = 500, err = -9999)
}
