% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dtmax.R
\name{calc_dtmax_pd_ed}
\alias{calc_dtmax_pd_ed}
\title{Calculate dTmax by the predawn method and environmental dependent method}
\usage{
calc_dtmax_pd_ed(
  vctr_time,
  vctr_dt,
  vctr_rs,
  vctr_vpd,
  thres_rs = 100,
  thres_vpd = 6,
  thres_cv = 0.005,
  thres_hour = 8,
  min_n_wndw = 3
)
}
\arguments{
\item{vctr_time}{Timestamp vector}

\item{vctr_dt}{Delta T (temperature difference between probes) time series}

\item{vctr_rs}{Incident short-wave radiation time series}

\item{vctr_vpd}{Water vapor pressure deficit time series}

\item{thres_rs}{Threshold value of predawn short-wave radiation}

\item{thres_vpd}{Threshold value of predawn short-wave radiation}

\item{thres_cv}{Threshold value of coefficient of variation}

\item{thres_hour}{Threshold value of predawn hour}

\item{min_n_wndw}{Threshold value of minimum data points in moving window}
}
\description{
calc_dtmax_pd_ed() calculates daily maximum temperature
  difference between granier-type sap flow probes (dTmax) by the predawn
  method and environmental dependent method.
}
\examples{
time_test <- dt_noisy$time
dt_test <- dt_noisy$dt
rs_test <- dt_noisy$rs
vpd_test <- dt_noisy$vpd

# Calculate dTmax from data
result <-
  calc_dtmax_pd_ed(time_test, dt_test, rs_test, vpd_test)
}
