% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxfixer-package.R
\docType{package}
\name{fluxfixer-package}
\alias{fluxfixer}
\alias{fluxfixer-package}
\title{fluxfixer: Advanced Framework for Sap Flow Data Post-Process}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a flexible framework for post-processing thermal dissipation sap flow data using statistical methods and machine learning. This framework includes anomaly correction, outlier removal, gap-filling, trend removal, signal damping correction, and sap flux density calculation. The functions in this package can also apply to other time series with various artifacts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yhata86/fluxfixer}
  \item \url{https://yhata86.github.io/fluxfixer/}
  \item Report bugs at \url{https://github.com/yhata86/fluxfixer/issues}
}

}
\author{
\strong{Maintainer}: Yoshiaki Hata \email{yoshiakihata0806@gmail.com} (\href{https://orcid.org/0000-0001-7703-0863}{ORCID}) [copyright holder]

}
\keyword{internal}
